refato(notification): mutualize checkboxSelection with simulation
