+++ gsl_core/templates/base.html
@@ -58,8 +58,7 @@
                     "bindAmountFields": "/static/js/controllers/bindAmountTields.js",
                     "formUtils": "/static/js/controllers/formUtils.js",
                     "dotationDropdown": "/static/js/controllers/dotationDropdown.js",
-                    "commentArbitrage": "/static/js/controllers/commentArbitrage.js",
-                    "bulkStatusChange": "/static/js/controllers/bulkStatusChange.js"
+                    "commentArbitrage": "/static/js/controllers/commentArbitrage.js"
                 }
             }
             </script>
@@ -132,7 +131,6 @@
             import { FormUtils } from "formUtils"
             import { DotationDropdown } from "dotationDropdown"
             import { CommentArbitrage } from "commentArbitrage"
-            import { BulkStatusChange } from "bulkStatusChange"
 
             window.Stimulus = Application.start()
 
@@ -146,7 +144,6 @@
             Stimulus.register("form-utils", FormUtils)
             Stimulus.register("dotation-dropdown", DotationDropdown)
             Stimulus.register("comment-arbitrage", CommentArbitrage)
-            Stimulus.register("bulk-status-change", BulkStatusChange)
         </script>
 
         {% block extra_js %}
