+++ gsl_simulation/static/css/simulation_detail.css
@@ -1,7 +1,3 @@
-.gsl-table-toolbar .fr-btns-group .fr-btn {
-  margin-bottom: 0;
-}
-
 /* table */
 .gsl-projet-table__select {
   text-align: right;
