+++ gsl_simulation/templates/includes/_simulation_bulk_actions_toolbar.html
@@ -12,17 +12,17 @@
 {% endcomment %}
 
 <div class="gsl-table-toolbar fr-mt-11v fr-mb-6v">
-    {{ selectable_ids_list|json_script:"bulk-status-change-selectable-ids" }}
+    {{ selectable_ids_list|json_script:"checkbox-selection-selectable-ids" }}
     <span class="gsl-bulk-actions__count">
         {{ page_obj.paginator.count }} projet{{ page_obj.paginator.count|custom_pluralize }}
         <span class="fr-text-mention--grey fr-ml-2w"
-              data-bulk-status-change-target="counterWrapper">
-            <span data-bulk-status-change-target="counter">0</span> projets sélectionnés
+              data-checkbox-selection-target="counterWrapper">
+            <span data-checkbox-selection-target="counter">0</span> projets sélectionnés
         </span>
     </span>
     <ul class="fr-btns-group fr-btns-group--inline fr-btns-group--right fr-btns-group--icon-right">
         <li>
-            <button type="button" class="fr-btn fr-btn--tertiary" data-bulk-status-change-target="selectAllButton" data-action="click->bulk-status-change#toggleSelectAll"
+            <button type="button" class="fr-btn fr-btn--tertiary" data-checkbox-selection-target="selectAllButton" data-action="click->checkbox-selection#toggleSelectAll"
                 {% if not selectable_count %}
                     disabled
                 {% endif %}
@@ -35,25 +35,25 @@
                 <div class="fr-nav__item">
                     <button type="button"
                             class="fr-btn fr-btn--secondary fr-btn--icon-right fr-icon-arrow-down-s-line"
-                            aria-controls="bulk-status-change-menu"
+                            aria-controls="checkbox-selection-bulk-menu"
                             aria-expanded="false"
-                            data-bulk-status-change-target="applyButton"
+                            data-checkbox-selection-target="actionButton"
                             disabled>
                         Modifier les statuts à
                     </button>
-                    <div class="fr-menu fr-collapse" id="bulk-status-change-menu">
+                    <div class="fr-menu fr-collapse" id="checkbox-selection-bulk-menu">
                         <ul class="fr-menu__list">
                             {% for status_key in bulk_status_choices %}
                                 <li>
                                     <form method="post"
                                           hx-post="{% url 'simulation:simulation-projet-bulk-update-simulation-status' status=status_key %}"
                                           hx-swap="none"
-                                          data-action="submit->bulk-status-change#beforeSubmit">
+                                          data-action="submit->checkbox-selection#beforeSubmit">
                                         {% csrf_token %}
                                         <input type="hidden"
                                                name="simulation_projet_ids"
                                                value=""
-                                               data-bulk-status-change-target="idsInput">
+                                               data-checkbox-selection-target="idsInput">
                                         <button type="submit" class="fr-nav__link">
                                             {{ status_key|status_to_label|split_symbol_and_status|safe }}
                                         </button>
